+++ application/views/UpdateStatus.php
@@ -1,0 +1,2 @@
+These series are scheduled to be updated in approximately: <?=$update_time[0]?> hours and <?=$update_time[1]?> minutes.
+<?=$this->table->generate($updates)?>